+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+* Add `Signal#cycle` method
+* Add `Signal#sequential` method
+* Deprecate `Signal.sequential` method (use `Signal.periodic(1000).sequential([1, 2, 3])` instead)
+
 ## 1.3.0 (2019-01-06)
 
 * Rename `emit.next` -> `emit.value`